Add progress into calculation

+++ app/src/main/java/com/christianv07/dev/speedy/model/DownloadTime.kt
@@ -3,7 +3,7 @@ package com.christianv07.dev.speedy.model
 data class DownloadTime(
     var fileSize: Double,
     var estimatedSpeed: Double,
-    var progress: Int,
+    var progress: Double,
     var fileSizeModifier: String,
     var estimatedSpeedModifier: String
 )
